+++ docs/releasenotes.md
@@ -4,6 +4,7 @@
 
 - add `[SimpleRpcExceptions]` (which can be applied at service contract, service type, or service method levels), and `SimpleRpcExceptionsInterceptor` (which can be applied to any service registration) - which expose server exceptions more conveniently (#75)
 - use linker-friendly metadata inspection (#90)
+- add non-generic `AddCodeFirst` native server overload (#106)
 
 ## 1.0.90